[K-S4-I] Update modlist and switch to API v3 (see desc)
API v2 sucks. I mean, it's not terrible, but the stuff it has is outdated nowadays, namely its lack of support for organizations. It also implements a boolean in project members requests called `is_owner`, which makes setting the owner for a project much easier. Because of this, even though API v3 isn't considered stable yet, we're switching to it instead :D

Projects owned by organizations also now will show the organization as the owner for the mod (yay!)