[K-S4-I] Speed up page load drastically
The reason it took so long to load originally was because for each mod a request was made to the API for the members/organization of the mod. Now, we make a bulk request for all of the data regarding the members of the mod and then pass that into `createMod()`, which passes it into `getProjectOwner()`. If a project is determined as being owned by an organization in `getProjectOwner()`, we still make an API request for the organization's data, because only a few mods are owned by organizations, so performance impact isn't that big of a deal. If we did make requests for organizations outside of `getProjectOwner()`, code complexity would increase to 1000 which wouldn't be worth it anyway.